refactor: consolidate health and release routes into a single meta routes endpoint
