--- src/test/kotlin/io/sdkman/broker/support/ReleaseTestSupport.kt
@@ -1,27 +1,0 @@
-package io.sdkman.broker.support
-
-import io.ktor.serialization.kotlinx.json.json
-import io.ktor.server.application.Application
-import io.ktor.server.application.install
-import io.ktor.server.plugins.contentnegotiation.ContentNegotiation
-import io.sdkman.broker.adapter.primary.rest.releaseRoutes
-import io.sdkman.broker.application.service.ReleaseService
-import kotlinx.serialization.ExperimentalSerializationApi
-import kotlinx.serialization.json.Json
-
-@OptIn(ExperimentalSerializationApi::class)
-fun Application.configureAppForReleaseTesting(releaseService: ReleaseService) {
-    install(ContentNegotiation) {
-        json(
-            Json {
-                prettyPrint = true
-                isLenient = true
-                ignoreUnknownKeys = true
-                encodeDefaults = true
-                explicitNulls = false
-            }
-        )
-    }
-
-    releaseRoutes(releaseService)
-}
